backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
requests
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
